+++ mail/internal/sesfake/sigv4_internal_test.go
@@ -1,4 +1,4 @@
-package mailtest
+package sesfake
 
 import (
 	"net/http"